requirements.txt
# Core Dependencies
python>=3.9

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0

# NLP and Transformers
transformers>=4.30.0
tokenizers>=0.13.0
sentence-transformers>=2.2.0

# LangChain and Orchestration
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Ollama Integration
ollama>=0.1.0
requests>=2.31.0

# Medical NLP Models
biobert-embedding>=0.1.2
spacy>=3.7.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Medical Imaging
SimpleITK>=2.2.0
pydicom>=2.4.0
nibabel>=5.1.0
scikit-image>=0.21.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Model Evaluation
scikit-metrics>=0.1.0
monai>=1.2.0  # Medical Open Network for AI

# Optional: For advanced medical imaging
# medpy>=0.4.0
# radiomics>=3.1.0
